% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dups-class.R
\name{as.data.frame.dups}
\alias{as.data.frame.dups}
\title{as.data.frame method for `dups` class}
\usage{
\method{as.data.frame}{dups}(x, ...)
}
\arguments{
\item{x}{any \R object.}

\item{...}{additional arguments to be passed to or from methods.}
}
\description{
as.data.frame method for `dups` class
}
