% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dups-class.R
\name{as_tibble.dups}
\alias{as_tibble.dups}
\title{convert a `dups` object to a `tibble`}
\usage{
\method{as_tibble}{dups}(x, ...)
}
\arguments{
\item{x}{A data frame, list, matrix, or other object that could reasonably be
coerced to a tibble.}

\item{...}{Other arguments passed on to individual methods.}
}
\description{
convert a `dups` object to a `tibble`
}
