% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dupree.R
\name{compare_code_blocks}
\alias{compare_code_blocks}
\title{Runs an alignment between all code blocks}
\usage{
compare_code_blocks(code_blocks, min_length = 20)
}
\arguments{
\item{code_blocks}{A vector of characters. If generated using the
standard workflow, each entry should correspond to a block of code in the
original R script and should have no whitespace or comments present.}

\item{min_length}{A non-negative integer. Only code-blocks that
have at least this many characters are used in the alignment step.}
}
\description{
For each block, it returns the index of the code block with the highest
similarity and the alignment score based on local-alignment using
pairwiseAlignment from Biostrings. Code blocks with fewer than `min_length`
characters are not included in the analysis
}
