% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dupree.R
\name{dupr}
\alias{dupr}
\title{Duplicate detection workflow}
\usage{
dupr(file, min_length = 20)
}
\arguments{
\item{file}{Some filename. Should be an R script or
R-markdown file (or similar file parseable by lintr).}

\item{min_length}{A non-negative integer. Only code-blocks that
have at least this many characters are used in the alignment step.}
}
\description{
Duplicate detection workflow
}
