% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dupree.R
\name{dupree_package}
\alias{dupree_package}
\title{`dupree_package` - run duplicate-code detection over all files in a
package's `R` directory}
\usage{
dupree_package(package, min_block_size = 20)
}
\arguments{
\item{package}{The name or path to the package that is to be
checked.}

\item{min_block_size}{`dupree` uses a notion of non-trivial
symbols.  These are the symbols / code-words that remain after filtering
out really common symbols like `<-`, `,`, etc. After filtering out these
symbols from each code-block, only those blocks containing at least
`min_block_size` symbols are used in the inter-block code-duplication
measurement.}
}
\description{
`dupree_package` - run duplicate-code detection over all files in a
package's `R` directory
}
