% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dupree_classes.R
\name{EnumeratedCodeTable-class}
\alias{EnumeratedCodeTable-class}
\title{An S4 class to represent the code blocks as strings of integers}
\description{
An S4 class to represent the code blocks as strings of integers
}
\section{Slots}{

\describe{
\item{\code{blocks}}{A tbl_df with columns `file`, `block`, `start_line` and
`enumerated_code`}
}}

