% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dupree_code_enumeration.R
\name{get_localised_parsed_code_blocks}
\alias{get_localised_parsed_code_blocks}
\title{Convert a list of source_expressions to a data-frame that contains the
  parsed-content from each source expression, and indicates the file,
  block-number and start-line for that source expression}
\usage{
get_localised_parsed_code_blocks(source_exprs)
}
\description{
Convert a list of source_expressions to a data-frame that contains the
  parsed-content from each source expression, and indicates the file,
  block-number and start-line for that source expression
}
