% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dupree.R
\name{parse_code_blocks}
\alias{parse_code_blocks}
\title{Extracts code-blocks from the source-expressions returned by lintr}
\usage{
parse_code_blocks(source_exprs)
}
\arguments{
\item{source_exprs}{Source-expressions data-structure as returned
by lintr::get_source_expressions.}
}
\description{
Extracts code-blocks from the source-expressions returned by lintr
}
