% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dups-class.R
\name{print.dups}
\alias{print.dups}
\title{print method for `dups` class}
\usage{
\method{print}{dups}(x, ...)
}
\arguments{
\item{x}{an object used to select a method.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
print method for `dups` class
}
