% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dupree_code_enumeration.R
\name{annotate_parsed_content}
\alias{annotate_parsed_content}
\title{Add filename, block-number and start-line for the parsed-content for each
  code block in a given file}
\usage{
annotate_parsed_content(parsed_content, file, block, start_line)
}
\description{
Add filename, block-number and start-line for the parsed-content for each
  code block in a given file
}
